Split `fragment.rs` into new module
